database is now updating our monthly and yealy too
